phishslayer-api/requirements.txt
# FILE: phishslayer-api/requirements.txt
# Python dependencies for PhishSlayer FastAPI backend

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async support
httpx==0.25.2
aiofiles==23.2.1

# Database
motor==3.3.2
pymongo==4.6.1
sqlalchemy==2.0.23

# Data processing
pandas==2.1.3
numpy==1.26.2

# ML/AI
openai==1.3.8
groq==0.4.1
langchain==0.1.0
agentops==0.3.0
agentscope

# Integrations
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0

# Security
cryptography==41.0.7
pyjwt==2.8.1
python-jose==3.3.0
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-testing==0.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Logging
python-json-logger==2.0.7
structlog==23.3.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1

pdfplumber
beautifulsoup4
scrapling
